refactor: add utils and helpers
