+++ src/app/cities/__utils__.ts
@@ -3,3 +3,15 @@ export enum CitySize {
   "Medium" = "🏙️ Middle-Sized City",
   "Big" = "🌆 Big City",
 }
+
+export enum Winter {
+  "Mild" = "🌨️ Mild Winter",
+  "Cold" = "❄️ Mild Winter",
+  "Freezing" = "🥶 Freezing Winter"
+}
+
+export enum Cost {
+  "Low" = "💸 Low",
+  "Medium" = "💸💸 Medium",
+  "High" = "💸💸💸 High"
+}
